refactor: move file panel rendering to render.lua
